pyproject.toml
[project]
name = "regstack"
version = "0.1.0.dev0"
description = "Embeddable user registration, login, and account management for FastAPI/MongoDB apps."
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
license-files = ["LICENSE", "NOTICE"]
authors = [{ name = "Joe Drumgoole", email = "joe@joedrumgoole.com" }]
keywords = ["fastapi", "mongodb", "auth", "registration", "users"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastapi>=0.110",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pymongo>=4.9",
    "pwdlib[argon2]>=0.2.1",
    "pyjwt>=2.8",
    "jinja2>=3.1",
    "click>=8.1",
    "dnspython>=2.6",
    "python-multipart>=0.0.9",
    "email-validator>=2.1",
    "aiosmtplib>=3.0",
]

[project.optional-dependencies]
ses = ["aioboto3>=12.3"]
sns = ["aioboto3>=12.3"]
twilio = ["twilio>=9.0"]
docs = [
    "sphinx>=7.3",
    "myst-parser>=3.0",
    "furo>=2024.1",
    "sphinx-autodoc-typehints>=2.0",
    "sphinx-copybutton>=0.5",
    "sphinx-autobuild>=2024.4",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-xdist>=3.5",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "anyio>=4.3",
    "ruff>=0.4",
    "mypy>=1.10",
    "invoke>=2.2",
    "uvicorn[standard]>=0.29",
]

[project.scripts]
regstack = "regstack.cli.__main__:main"

[project.urls]
Homepage = "https://github.com/jdrumgoole/regstack"
Repository = "https://github.com/jdrumgoole/regstack"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/regstack"]
# Hatch's package autodiscovery already picks up template / static assets
# under src/regstack — no force-include needed (it would double-pack).

[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "e2e: end-to-end Playwright tests (run serially, slow)",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:passlib.*",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "RUF"]
ignore = [
    "E501",
    # `Depends(...)` in argument defaults is the canonical FastAPI pattern.
    "B008",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
plugins = ["pydantic.mypy"]
